+++ Nickel.UpdateChecks/ModEntry.cs
@@ -58,7 +58,7 @@ public sealed class ModEntry : SimpleMod
 	private (Task Task, CancellationTokenSource Token)? CurrentUpdateCheckTask;
 
 	private IWritableFileInfo SettingsFile
-		=> this.Helper.Storage.GetSingleSettingsFile("json");
+		=> this.Helper.Storage.GetMainStorageFile("json");
 
 	public ModEntry(IPluginPackage<IModManifest> package, IModHelper helper, ILogger logger) : base(package, helper, logger)
 	{